requirement.txt
Django
Psycopg2
Psycopg2-binary
Djangorestframework
Djangorestframework-simplejwt
Djoser
pylint-django

npm install --save axios react-router-dom redux redux-devtools-extension react-redux redux-thumk